+++ plugins/plugin-md-power/tsup.config.ts
@@ -3,7 +3,7 @@ import { argv } from '../../scripts/tsup-args.js'
 
 const config = [
   { dir: 'composables', files: ['codeRepl.ts', 'pdf.ts', 'rustRepl.ts', 'size.ts', 'audio.ts'] },
-  { dir: 'utils', files: ['http.ts', 'is.ts', 'link.ts', 'sleep.ts'] },
+  { dir: 'utils', files: ['http.ts', 'is.ts', 'link.ts', 'sleep.ts', 'shared.ts'] },
   { dir: '', files: ['index.ts', 'options.ts'] },
 ]